Add author and category attributes to booksSlice.js
