+++ src/redux/features/books/booksSlice.js
@@ -1,22 +1,22 @@
 import { createSlice } from '@reduxjs/toolkit';
+import { v4 as uuidv4 } from 'uuid';
 
 const initialState = {
-  books:
-  [
+  books: [
     {
-      itemId: 'item1',
+      itemId: uuidv4(),
       title: 'The Great Gatsby',
       author: 'John Smith',
       category: 'Fiction',
     },
     {
-      itemId: 'item2',
+      itemId: uuidv4(),
       title: 'Anna Karenina',
       author: 'Leo Tolstoy',
       category: 'Fiction',
     },
     {
-      itemId: 'item3',
+      itemId: uuidv4(),
       title: 'The Selfish Gene',
       author: 'Richard Dawkins',
       category: 'Nonfiction',
@@ -29,7 +29,13 @@ const booksSlice = createSlice({
   initialState,
   reducers: {
     addBook: (state, action) => {
-      state.books.push(action.payload);
+      const newBook = {
+        itemId: uuidv4(),
+        title: action.payload.title,
+        author: action.payload.author,
+        category: action.payload.category,
+      };
+      state.books.push(newBook);
     },
     removeBook: (state, action) => {
       state.books = state.books.filter((book) => book.itemId !== action.payload);
